+++ src/components/load-more.js
@@ -1,0 +1,5 @@
+export const getLoad = () => {
+  return `
+    <button class="load-more" type="button">load more</button>
+  `;
+};
