+++ src/components/menu-wrapper.js
@@ -1,0 +1,9 @@
+export const getMenuWrapper = (menuItem) => {
+  return `
+    <section class="control__btn-wrap">
+      ${menuItem(`new-task`, `+ADD NEW TASK`)}
+      ${menuItem(`task`, `TASKS`)}
+      ${menuItem(`statistics`, `STATISTICS`)}
+    </section>
+  `;
+};
