+++ src/utils.js
@@ -1,0 +1,3 @@
+const getRandom = (min, max) => Math.floor(min + Math.random() * (max + 1 - min));
+
+export {getRandom};
